feat: separate top bar into desktop and mobile
